server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.0.0
jsonschema>=4.17.0
openapi-spec-validator>=0.6.0
orjson>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0

